+++ Stack/Stack/Stack.h
@@ -4,7 +4,7 @@
 // Structure for implementing a stack consisting of a value and a pointer to the next element
 typedef struct Stack
 {
-    int value;
+    float value;
     struct Stack* next;
 } Stack;
 
@@ -12,10 +12,10 @@ typedef struct Stack
 bool isEmpty(Stack* head);
 
 // Function for adding an element to the top of the stack
-void push(Stack** head, int element);
+void push(Stack** head, float element);
 
 // Function to remove an element from the top of the stack that returns the value of that element
-int pop(Stack** head);
+float pop(Stack** head);
 
 // Function for deleting all stack elements
 void deleteStack(Stack** head);
