+++ Stack/Stack/StackTest.c
@@ -6,9 +6,9 @@ bool pushTest()
 {
     Stack* head = NULL;
     push(&head, 12);
-    int firstResult = head->value;
+    float firstResult = head->value;
     push(&head, 128);
-    int secondResult = head->value;
+    float secondResult = head->value;
     return firstResult == 12 && secondResult == 128;
 }
 
@@ -18,10 +18,10 @@ bool popTest()
     push(&head, 12);
     push(&head, 128);
     push(&head, 147);
-    int firstPopResult = pop(&head);
-    int firstUpperElement = head->value;
-    int secondPopResult = pop(&head);
-    int secondUpperElement = head->value;
+    float firstPopResult = pop(&head);
+    float firstUpperElement = head->value;
+    float secondPopResult = pop(&head);
+    float secondUpperElement = head->value;
     return firstPopResult == 147 && firstUpperElement == 128 && secondPopResult == 128 && secondUpperElement == 12;
 }
